Add files via upload
When it comes to cooking, everyone has a specialty. Create a recipe, with at least five ingredients You can assume that:

All ingredients are available and accessible to the user.
The person reading the recipe will know how to measure, pour, etc.
The person is literal.
There is no alternative to ingredients
You need to allow for a change in the number of people (i.e. your recipe may be for 4 people, but the person reading the recipe may need to double or triple it. Your program should ask for a number of people and compute the multiplying factor).
When the user runs the program, the recipe will be printed, with all necessary steps from start to finish.
You will need to have a minimum of five ingredients.

Type the answers in a word document or an open offic